delete backend as subdir and add to lib with renaming to math
