+++ CMakeLists.txt
@@ -45,7 +45,6 @@ add_subdirectory(${QCUSTOMPLOT_DIR}) # here adding qcustomplot subfolder
 add_subdirectory(lib)
 add_subdirectory(tests)
 add_subdirectory(main)
-add_subdirectory(backend)
 
 # --------------------------------------------------------------------
