+++ math/adjacency_matrix.cpp
