+++ math/adjacency_matrix.h
