+++ math/littles_algorithm/adjacency_matrix.h
