+++ math/optimal_way/circle_obstacle.h
