+++ math/optimal_way/helpers_functions.cpp
