+++ math/optimal_way/optimal_way.cpp
