+++ math/optimal_way/path_graph.cpp
