+++ math/optimal_way/path_graph.h
