+++ tests/CMakeLists.txt
@@ -34,7 +34,6 @@ endif()
 target_link_libraries(tests PUBLIC Qt${QT_VERSION_MAJOR}::Core Qt${QT_VERSION_MAJOR}::Gui Qt${QT_VERSION_MAJOR}::Widgets Qt${QT_VERSION_MAJOR}::PrintSupport)
 
 target_link_libraries(tests PRIVATE lib)
-target_link_libraries(tests PRIVATE backend)
 
 target_link_libraries(tests PRIVATE ${Boost_LIBRARIES})
