+++ tests/maths/many_tsp_tests.cpp
@@ -1,8 +1,8 @@
 #include <QApplication>
 
-#include "../backend/littles_algorithm/adjacency_matrix.h"
-#include "../backend/littles_algorithm/travelling_salesmans_problem.h"
-#include "../backend/littles_algorithm/tspgraph.h"
+#include "../math/littles_algorithm/adjacency_matrix.h"
+#include "../math/littles_algorithm/travelling_salesmans_problem.h"
+#include "../math/littles_algorithm/tspgraph.h"
 
 #if !defined(WIN32)
 #define BOOST_TEST_DYN_LINK
